+++ src/data_reliability/__init__.py
@@ -1,15 +1,30 @@
 from .core import DataTier, ReliableData, ReliabilityMetadata, ReliabilityPolicy
 from .pandas_ext import filter_reliable_df
-from .scanner import ReliabilityScanner, ReliabilityWeights, ValidationEvidence, compute_trace_hash
+from .scanner import (
+    ReliabilityProfile,
+    ReliabilityScanner,
+    ReliabilityWeights,
+    TierCriterion,
+    ValidationEvidence,
+    climate_record_profile,
+    compute_trace_hash,
+    default_profile,
+    scientific_profile,
+)
 
 __all__ = [
     "DataTier",
     "ReliableData",
     "ReliabilityMetadata",
     "ReliabilityPolicy",
+    "ReliabilityProfile",
     "ReliabilityScanner",
     "ReliabilityWeights",
+    "TierCriterion",
     "ValidationEvidence",
+    "climate_record_profile",
     "compute_trace_hash",
+    "default_profile",
     "filter_reliable_df",
+    "scientific_profile",
 ]